requirements.txt
selene==2.0.0rc4
pytest==7.4.3
allure-pytest==2.9.45